Update type hints

+++ tafrigh/transcript_writer.py
@@ -8,7 +8,7 @@ class TranscriptWriter:
         self,
         format: TranscriptType,
         file_path: str,
-        segments: List[Dict[str, Union[str, int]]],
+        segments: List[Dict[str, Union[str, float]]],
         min_words_per_segment: int = 0,
     ) -> None:
         if format == TranscriptType.VTT:
@@ -21,7 +21,7 @@ class TranscriptWriter:
     def write_vtt(
         self,
         file_path: str,
-        segments: List[Dict[str, Union[str, int]]],
+        segments: List[Dict[str, Union[str, float]]],
         min_words_per_segment: int = 0,
     ) -> None:
         segments = self._compact_segments(segments, min_words_per_segment)
@@ -30,7 +30,7 @@ class TranscriptWriter:
     def write_srt(
         self,
         file_path: str,
-        segments: List[Dict[str, Union[str, int]]],
+        segments: List[Dict[str, Union[str, float]]],
         min_words_per_segment: int = 0,
     ) -> None:
         segments = self._compact_segments(segments, min_words_per_segment)
@@ -39,20 +39,20 @@ class TranscriptWriter:
     def write_txt(
         self,
         file_path: str,
-        segments: List[Dict[str, Union[str, int]]],
+        segments: List[Dict[str, Union[str, float]]],
         min_words_per_segment: int = 0,
     ) -> None:
         segments = self._compact_segments(segments, min_words_per_segment)
         self._write_to_file(file_path, self.generate_txt(segments))
 
-    def generate_vtt(self, segments: List[Dict[str, Union[str, int]]]) -> str:
+    def generate_vtt(self, segments: List[Dict[str, Union[str, float]]]) -> str:
         return 'WEBVTT\n\n' + ''.join(
             f"{self._format_timestamp(segment['start'])} --> {self._format_timestamp(segment['end'])}\n"
             f"{segment['text'].strip()}\n\n"
             for segment in segments
         )
 
-    def generate_srt(self, segments: List[Dict[str, Union[str, int]]]) -> str:
+    def generate_srt(self, segments: List[Dict[str, Union[str, float]]]) -> str:
         return ''.join(
             f"{i}\n"
             f"{self._format_timestamp(segment['start'], include_hours=True, decimal_marker=',')} --> "
@@ -61,7 +61,7 @@ class TranscriptWriter:
             for i, segment in enumerate(segments, start=1)
         )
 
-    def generate_txt(self, segments: List[Dict[str, Union[str, int]]]) -> str:
+    def generate_txt(self, segments: List[Dict[str, Union[str, float]]]) -> str:
         return '\n'.join(list(map(lambda segment: segment['text'].strip(), segments))) + '\n'
 
     def _write_to_file(self, file_path: str, content: str) -> None:
@@ -86,9 +86,9 @@ class TranscriptWriter:
 
     def _compact_segments(
         self,
-        segments: List[Dict[str, Union[str, int]]],
+        segments: List[Dict[str, Union[str, float]]],
         min_words_per_segment: int,
-    ) -> List[Dict[str, Union[str, int]]]:
+    ) -> List[Dict[str, Union[str, float]]]:
         if min_words_per_segment == 0:
             return segments
